third_party/blink/web_tests/external/wpt/html/rendering/non-replaced-elements/tables/table-attribute-expected.txt

This is a testharness.js-based test.
Found 24 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] table thead align attribute center is correct
  assert_equals: expected "center" but got "-webkit-center"
[FAIL] table thead align attribute middle is correct
  assert_equals: expected "center" but got "-webkit-center"
[FAIL] table thead align attribute left is correct
  assert_equals: expected "left" but got "-webkit-left"
[FAIL] table thead align attribute right is correct
  assert_equals: expected "right" but got "-webkit-right"
[FAIL] table tbody align attribute center is correct
  assert_equals: expected "center" but got "-webkit-center"
[FAIL] table tbody align attribute middle is correct
  assert_equals: expected "center" but got "-webkit-center"
[FAIL] table tbody align attribute left is correct
  assert_equals: expected "left" but got "-webkit-left"
[FAIL] table tbody align attribute right is correct
  assert_equals: expected "right" but got "-webkit-right"
[FAIL] table tfoot align attribute center is correct
  assert_equals: expected "center" but got "-webkit-center"
[FAIL] table tfoot align attribute middle is correct
  assert_equals: expected "center" but got "-webkit-center"
[FAIL] table tfoot align attribute left is correct
  assert_equals: expected "left" but got "-webkit-left"
[FAIL] table tfoot align attribute right is correct
  assert_equals: expected "right" but got "-webkit-right"
[FAIL] table tr align attribute center is correct
  assert_equals: expected "center" but got "-webkit-center"
[FAIL] table tr align attribute middle is correct
  assert_equals: expected "center" but got "-webkit-center"
[FAIL] table tr align attribute left is correct
  assert_equals: expected "left" but got "-webkit-left"
[FAIL] table tr align attribute right is correct
  assert_equals: expected "right" but got "-webkit-right"
[FAIL] table td align attribute center is correct
  assert_equals: expected "center" but got "-webkit-center"
[FAIL] table td align attribute middle is correct
  assert_equals: expected "center" but got "-webkit-center"
[FAIL] table td align attribute left is correct
  assert_equals: expected "left" but got "-webkit-left"
[FAIL] table td align attribute right is correct
  assert_equals: expected "right" but got "-webkit-right"
[FAIL] table th align attribute center is correct
  assert_equals: expected "center" but got "-webkit-center"
[FAIL] table th align attribute middle is correct
  assert_equals: expected "center" but got "-webkit-center"
[FAIL] table th align attribute left is correct
  assert_equals: expected "left" but got "-webkit-left"
[FAIL] table th align attribute right is correct
  assert_equals: expected "right" but got "-webkit-right"
Harness: the test ran to completion.
